# docs/swagger.yaml
definitions:
...
  types.ApiResponse:
    properties:
      data: {}
      status:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/eth1deposit/{txhash}:
    get:
      parameters:
      - description: Eth1 transaction hash
        in: path
        name: txhash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiResponse'
      summary: Get an eth1 deposit by its eth1 transaction hash
      tags:
      - Execution
...
swagger: "2.0"